move DropDown and enable absolute imports
