+++ tsconfig.json
@@ -18,7 +18,8 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "react",
-    "sourceMap": true
+    "sourceMap": true,
+    "baseUrl": "src"
   },
   "include": [
     "src"
